Move ActionApplication.ts and OptimisticJob to action/
